style bugfix on years input and separate compounding logic file
