+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "dx_fintools_fs"
 version = "0.1.0"
-authors = ["aaelony <aaelony@gmail.com>"]
+authors = ["aaelony"]
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
